I seem to have finally figured out movement.
Now we use a RigidBody2D in Character mode with any sort of engine-calculated physics disabled. Then we use the test_motion method of the rigid body and the rest is more-or-less the same. One strange issue I found was that when implementing the kicking system, pushing the tetromino down would sometimes cause it to glitch out but side kicks and ground kicks seems to work fine (for now).